refactor: extract grid logic to a separate mixin
